Feat/NaverLogin #7 에러 고침. 아이콘 원형으로
